Add core conversion logic, some argument parsing
At this point, I've put in the basic conversion and display components.
I've also added some of the logic for the argument parsing, but I'm
still experimenting with that, so the final work will look a bit
different.  I should've put in a commit without that stuff, but here it
is.